+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - plugin panel fixation when switching tabs
+- recovery dashboard state from URL
 
 ## [0.15.0]